+++ backend/src/main/java/com/pickpick/entity/Message.java
@@ -1,6 +1,7 @@
 package com.pickpick.entity;
 
 import com.pickpick.channel.domain.Channel;
+import com.pickpick.member.domain.Member;
 import java.time.LocalDateTime;
 import javax.persistence.Column;
 import javax.persistence.Entity;
