+++ backend/src/main/java/com/pickpick/member/domain/MemberRepository.java
@@ -1,6 +1,5 @@
-package com.pickpick.repository;
+package com.pickpick.member.domain;
 
-import com.pickpick.entity.Member;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.repository.Repository;
