+++ backend/src/main/java/com/pickpick/service/MessageCreatedService.java
@@ -4,10 +4,10 @@ import com.pickpick.channel.domain.Channel;
 import com.pickpick.channel.domain.ChannelRepository;
 import com.pickpick.controller.dto.MessageDto;
 import com.pickpick.controller.event.SlackEvent;
-import com.pickpick.entity.Member;
 import com.pickpick.exception.ChannelNotFoundException;
 import com.pickpick.exception.MemberNotFoundException;
-import com.pickpick.repository.MemberRepository;
+import com.pickpick.member.domain.Member;
+import com.pickpick.member.domain.MemberRepository;
 import com.pickpick.repository.MessageRepository;
 import java.util.Map;
 import org.springframework.stereotype.Service;
